==> 02_Traffic_Detection-No_Streaming.ipynb <==
__Libraries__

from pyspark import SparkContext
from pyspark.sql import SparkSession, Row
from pyspark.sql.types import StructType
from pyspark.sql.functions import monotonically_increasing_id, row_number
from pyspark.sql.window import Window

from geopy.geocoders import Nominatim
import geopy
from geopy.extra.rate_limiter import RateLimiter
geopy.geocoders.options.default_user_agent = "BDBA"
from tqdm import tqdm

__Open Environment__

sc = SparkContext(appName="SparkforEachRDDapp")
spark = SparkSession(sc)

__Read Data__

vehic= spark.read.json('vehicles.json')
vehic.show()

roads= spark.read.json('roads.json')
roads.show()

**Get Road Names**

def road_from_coord(lat, lon):
    coordinates= str(lat)+ ','+str(lon)
    #10 mins maximum timeout to prevent being blocked
    locator = Nominatim(timeout=10)
    rgeocode = RateLimiter(locator.reverse, min_delay_seconds=0.001)
    location = rgeocode(coordinates)
    return(location.raw['address']['road'])

road_names = []
for i in tqdm(range(vehic.count())):
    road_names.append(road_from_coord(vehic.collect()[i]['latitude'],vehic.collect()[i]['longitude']))

__Add them to the df__

row = Row("road_name")
rdd = spark.sparkContext.parallelize(road_names)
rf=rdd.map(row).toDF()

vehic=vehic.withColumn('row_index', row_number().over(Window.orderBy(monotonically_increasing_id())))
rf=rf.withColumn('row_index', row_number().over(Window.orderBy(monotonically_increasing_id())))

vehic = vehic.join(rf, ["row_index"],how="inner").drop("row_index")

vehic.show()

**Join Both Feeds**

df=vehic.join(roads, ['road_name'], how="inner")

df.show()

**Define Traffic Location**

If accelerometer < min_speed and count(vehicle_id) > (50*number_lanes) we declare there is traffic

df.createOrReplaceTempView("df")
results = spark.sql("select road_name, count(vehicle_id) from df \
                    where accelerometer < min_speed \
                    group by road_name, number_lanes, has_bus_lane\
                    having count(vehicle_id) > (50*(number_lanes-has_bus_lane))")
results.show()

__Download Results__

https://community.cloudera.com/t5/Support-Questions/How-to-save-all-the-output-of-pyspark-sql-query-into-a-text/td-p/204560

#results.coalesce(1).write.csv("results_6000.csv")
results.write.csv("td_results_6000.csv")

__Open Results__

import pandas as pd
myres = pd.read_csv("td_results_6000.csv", names=['road_name', 'number_vehicles'])

myres
