==> 03_Traffic_Detection-Streaming.ipynb <==
__Libraries__

from pyspark import SparkContext
from pyspark.sql import SparkSession, Row, SQLContext
from pyspark.streaming import StreamingContext
from pyspark.sql.types import StructType
from pyspark.sql.functions import monotonically_increasing_id, row_number
from pyspark.sql.window import Window

from geopy.geocoders import Nominatim
import geopy
from geopy.extra.rate_limiter import RateLimiter
geopy.geocoders.options.default_user_agent = "BDBA"
from tqdm import tqdm

__Functions__

def road_from_coord(vehicle):
    lat = vehicle.select(vehicle['latitude'])
    lon = vehicle.select(vehicle['longitude'])
    coordinates= str(lat)+ ','+str(lon)
    #10 mins maximum timeout to prevent being blocked
    locator = Nominatim(timeout=10)
    rgeocode = RateLimiter(locator.reverse, min_delay_seconds=0.001)
    location = rgeocode(coordinates)
    return(location.raw['address']['road'])

__Open Context__

sc = SparkContext(appName="SparkforEachRDDapp")
spark = SparkSession(sc)
sqlContext = SQLContext(sc)
ssc = StreamingContext(sc, 10)

__Read Data__

inputPath = "./vehicdir"
df=spark.readStream.format('json').schema( StructType()
                .add("vehicle_id", "string")
                .add("datetime_utc", "float")
                .add("latitude", "float")
                .add("longitude", "float") 
                .add("accelerometer", "integer")
                .add("vehicle_type", 'string')
             ).load(inputPath)

print(df)

__Transform Data__

query1 = df.writeStream.foreach(road_from_coord).queryName('getroads').start()
query1.status

df.createOrReplaceTempView("df")
spark.sql("select * from df").show(1000, truncate=False)

query2 = df.writeStream.foreach(lambda rdd: road_from_coord(rdd)).start()
query2.status

query3 = df.foreachRDD(lambda rdd: road_from_coord(rdd))
query3.status

### Other Things Tested

__Read Data 1__

inputdir = "/work/students/Project Cars (Camila, Paula, Juan, Rocio, Oriol, Valeria)/Project/roadsdir/roads.json"

lines = spark.readStream.format('json').schema(StructType()
               .add("road_name", "string", True)
               .add("number_lanes", "integer")
               .add("min_speed", "integer")
               .add("has_bus_lane", "string")).option('path', "./roadsdir/roads.json").load()

import json

lines = ssc.textFileStream("data/roads.json") # process files as they appear
data = lines.map(json.loads) # map DStream and return new DStream

lines.pprint()

data.pprint()

**Read Data 2**

road_schema = ( StructType()
               .add("road_name", "string", True)
               .add("number_lanes", "integer")
               .add("min_speed", "integer")
               .add("has_bus_lane", "string")
             )
vehic_schema = ( StructType()
                .add("vehicle_id", "string")
                .add("datetime_utc", "float")
                .add("latitude", "float")
                .add("longitude", "float") 
                .add("accelerometer", "integer")
                .add("vehicle_type", 'string')
             )


roadstream = spark.readStream.json('./roadsdir/roads.json', schema=road_schema)
vehicstream = spark.readStream.json('./vehicdir/vehicles.json', schema=vehic_schema)

display(vehicstream)

vehicstream.select(vehicstream['latitude'], vehicstream['longitude']).show()

__Read Data 3__

roads = spark.read.json('roads.json')
sc.parallelize(roads)
